feat: change class to more close module
